+++ Libplanet/Store/StoreExtension.cs
@@ -58,21 +58,50 @@ namespace Libplanet.Store
         /// </summary>
         /// <param name="store">A store object.</param>
         /// <param name="namespace">The namespace to look up state references.</param>
+        /// <param name="onlyAfter">Includes state references only made after the block
+        /// this argument refers to, if present.</param>
+        /// <param name="ignoreAfter">Excludes state references made after the block
+        /// this argument refers to, if present.</param>
         /// <returns>A dictionary of account addresses to lists of their corresponding state
         /// references.  Each list of state references is in ascending order, i.e., the block
         /// closest to the genesis goes first and the block closest to the tip goes last.</returns>
         public static IImmutableDictionary<Address, IImmutableList<HashDigest<SHA256>>>
-        ListAllStateReferences(this IStore store, string @namespace)
+        ListAllStateReferences(
+            this IStore store,
+            string @namespace,
+            HashDigest<SHA256>? onlyAfter = null,
+            HashDigest<SHA256>? ignoreAfter = null
+        )
         {
+            (HashDigest<SHA256>, long)? baseBlock =
+                onlyAfter is HashDigest<SHA256> @base && store.GetBlockIndex(@base) is long baseIdx
+                    ? (@base, baseIdx)
+                    : null as (HashDigest<SHA256>, long)?;
+            (HashDigest<SHA256>, long)? targetBlock =
+                ignoreAfter is HashDigest<SHA256> tgt && store.GetBlockIndex(tgt) is long tgtIdx
+                    ? (tgt, tgtIdx)
+                    : null as (HashDigest<SHA256>, long)?;
             return store.ListAddresses(@namespace).Select(address =>
             {
-                ImmutableList<HashDigest<SHA256>> refs = store
-                    .IterateStateReferences(@namespace, address)
+                IEnumerable<Tuple<HashDigest<SHA256>, long>> refIndices =
+                    store.IterateStateReferences(@namespace, address);
+
+                if (targetBlock is ValueTuple<HashDigest<SHA256>, long> targetIndex)
+                {
+                    refIndices = refIndices.SkipWhile(p => p.Item2 > targetIndex.Item2);
+                }
+
+                if (baseBlock is ValueTuple<HashDigest<SHA256>, long> baseIndex)
+                {
+                    refIndices = refIndices.TakeWhile(p => p.Item2 >= baseIndex.Item2);
+                }
+
+                ImmutableList<HashDigest<SHA256>> refs = refIndices
                     .Select(p => p.Item1)
                     .Reverse()
                     .ToImmutableList();
                 return new KeyValuePair<Address, IImmutableList<HashDigest<SHA256>>>(address, refs);
-            }).ToImmutableDictionary();
+            }).Where(pair => pair.Value.Any()).ToImmutableDictionary();
         }
     }
 }